+++ crates/melody_cli/src/utils.rs
@@ -1,3 +1,5 @@
+use crate::CliError;
+use std::fs::write;
 use std::io::{self, Write};
 
 pub fn read_input() -> io::Result<String> {
@@ -18,4 +20,14 @@ pub fn exit(code: ExitCode) {
         ExitCode::Ok => std::process::exit(0),
         ExitCode::Error => std::process::exit(1),
     }
+}
+
+pub fn write_output_to_file(
+    output_file_path: String,
+    compiler_output: String,
+) -> Result<(), CliError> {
+    write(&output_file_path, compiler_output)
+        .map_err(|_| CliError::WriteFileError(output_file_path))?;
+
+    Ok(())
 }