+++ crates/melody_compiler/src/lib.rs
@@ -1,3 +1,5 @@
+#![forbid(unsafe_code)]
+
 mod ast;
 pub mod errors;
 mod source;
@@ -8,6 +10,22 @@ use errors::ParseError;
 use source::to_source;
 use utils::format_line_comments;
 
+/**
+Compiles Melody source code to a regular expression.
+
+see also: [`ParseError`]
+
+# Example
+
+```rust
+use melody_compiler::compiler;
+
+let source = r#"1 to 5 of "A";"#;
+let output = compiler(source);
+
+assert_eq!(output.unwrap(), "A{1,5}");
+```
+*/
 pub fn compiler(source: &str) -> Result<String, ParseError> {
     let formatted_source = format_line_comments(source);
 
@@ -15,5 +33,5 @@ pub fn compiler(source: &str) -> Result<String, ParseError> {
 
     let output = to_source(&ast);
 
-    Ok(format!("/{output}/"))
+    Ok(output)
 }