+++ crates/melody_compiler/tests/mod.rs
@@ -8,7 +8,7 @@ fn quantifier_test() {
   5 of "A";
   "#,
     );
-    assert_eq!(output.unwrap(), "/A{5}/");
+    assert_eq!(output.unwrap(), "A{5}");
 }
 
 #[test]
@@ -21,7 +21,7 @@ fn capture_test() {
       }
       "#,
     );
-    assert_eq!(output.unwrap(), "/(A{5}[0-9])/");
+    assert_eq!(output.unwrap(), "(A{5}[0-9])");
 }
 
 #[test]
@@ -34,7 +34,7 @@ fn named_capture_test() {
       }
       "#,
     );
-    assert_eq!(output.unwrap(), "/(?<name>A{5}[0-9])/");
+    assert_eq!(output.unwrap(), "(?<name>A{5}[0-9])");
 }
 
 #[test]
@@ -44,7 +44,7 @@ fn number_quantifier_range_test() {
       1 to 5 of "A";
       "#,
     );
-    assert_eq!(output.unwrap(), "/A{1,5}/");
+    assert_eq!(output.unwrap(), "A{1,5}");
 }
 
 #[test]
@@ -55,7 +55,7 @@ fn uppercase_range_test() {
       7 of A to Z;
       "#,
     );
-    assert_eq!(output.unwrap(), "/[A-Z](?:[A-Z]){7}/");
+    assert_eq!(output.unwrap(), "[A-Z](?:[A-Z]){7}");
 }
 
 #[test]
@@ -66,7 +66,7 @@ fn lowercase_range_test() {
       8 of a to z;
       "#,
     );
-    assert_eq!(output.unwrap(), "/[a-z](?:[a-z]){8}/");
+    assert_eq!(output.unwrap(), "[a-z](?:[a-z]){8}");
 }
 
 #[test]
@@ -76,7 +76,7 @@ fn open_range_expression_test() {
       over 4 of "a";
       "#,
     );
-    assert_eq!(output.unwrap(), "/a{5,}/");
+    assert_eq!(output.unwrap(), "a{5,}");
 }
 
 #[test]
@@ -88,7 +88,7 @@ fn start_end_test() {
       <end>;
       "#,
     );
-    assert_eq!(output.unwrap(), "/^a$/");
+    assert_eq!(output.unwrap(), "^a$");
 }
 
 #[test]
@@ -118,7 +118,7 @@ fn symbol_test() {
     );
     assert_eq!(
         output.unwrap(),
-        r"/^.\s\S\n\t\r\f\0\d\D\w\W\v[a-zA-Z] $\b[\b]/"
+        r"^.\s\S\n\t\r\f\0\d\D\w\W\v[a-zA-Z] $\b[\b]"
     );
 }
 
@@ -131,7 +131,7 @@ fn match_test() {
   0 to 9;
 }"#,
     );
-    assert_eq!(output.unwrap(), "/(?:A{5}[0-9]){3}/");
+    assert_eq!(output.unwrap(), "(?:A{5}[0-9]){3}");
 }
 
 #[test]
@@ -145,7 +145,7 @@ fn comment_test() {
     } 
     "#,
     );
-    assert_eq!(output.unwrap(), "/[0-5](?:x)/");
+    assert_eq!(output.unwrap(), "[0-5](?:x)");
 }
 
 #[test]
@@ -155,7 +155,7 @@ fn char_test() {
       3 of <char>;
       "#,
     );
-    assert_eq!(output.unwrap(), "/.{3}/");
+    assert_eq!(output.unwrap(), ".{3}");
 }
 
 #[test]
@@ -165,7 +165,7 @@ fn negative_range_test() {
       not 3 to 5;
       "#,
     );
-    assert_eq!(output.unwrap(), "/[^3-5]/");
+    assert_eq!(output.unwrap(), "[^3-5]");
 }
 
 #[test]
@@ -175,13 +175,13 @@ fn some_test() {
       some of <char>;
       "#,
     );
-    assert_eq!(single_output.unwrap(), "/.+/");
+    assert_eq!(single_output.unwrap(), ".+");
     let multiple_output = compiler(
         r#"
       some of "ABC";
       "#,
     );
-    assert_eq!(multiple_output.unwrap(), "/(?:ABC)+/");
+    assert_eq!(multiple_output.unwrap(), "(?:ABC)+");
 }
 
 #[test]
@@ -191,13 +191,13 @@ fn option_test() {
       option of <char>;
       "#,
     );
-    assert_eq!(single_output.unwrap(), "/.?/");
+    assert_eq!(single_output.unwrap(), ".?");
     let multiple_output = compiler(
         r#"
       option of "ABC";
       "#,
     );
-    assert_eq!(multiple_output.unwrap(), "/(?:ABC)?/");
+    assert_eq!(multiple_output.unwrap(), "(?:ABC)?");
 }
 
 #[test]
@@ -215,7 +215,7 @@ fn either_test() {
       }
       "#,
     );
-    assert_eq!(output.unwrap(), "/(?:first|second|[a-z])(?:first|second)/");
+    assert_eq!(output.unwrap(), "(?:first|second|[a-z])(?:first|second)");
 }
 
 #[test]
@@ -225,13 +225,13 @@ fn any_test() {
       any of <char>;
       "#,
     );
-    assert_eq!(single_output.unwrap(), "/.*/");
+    assert_eq!(single_output.unwrap(), ".*");
     let multiple_output = compiler(
         r#"
         any of "ABC";
       "#,
     );
-    assert_eq!(multiple_output.unwrap(), "/(?:ABC)*/");
+    assert_eq!(multiple_output.unwrap(), "(?:ABC)*");
 }
 
 #[test]
@@ -241,7 +241,7 @@ fn raw_test() {
       5 of `.*`;
       "#,
     );
-    assert_eq!(output.unwrap(), "/(?:.*){5}/");
+    assert_eq!(output.unwrap(), "(?:.*){5}");
 }
 
 #[test]
@@ -262,7 +262,7 @@ fn assertion_test() {
       }
       "#,
     );
-    assert_eq!(output.unwrap(), "/(?=a)(?<=a)(?!a)(?<!a)/");
+    assert_eq!(output.unwrap(), "(?=a)(?<=a)(?!a)(?<!a)");
 }
 
 #[test]
@@ -273,5 +273,5 @@ fn negative_char_class_test() {
         5 of not abcd;
         "#,
     );
-    assert_eq!(output.unwrap(), "/[^abcd](?:[^abcd]){5}/");
+    assert_eq!(output.unwrap(), "[^abcd](?:[^abcd]){5}");
 }